Added slight delay to further avoid rate limit issues.
Instead of just sleeping for a single second, the enforce_call_rate thread now waits for an additional 40 ms to avoid varying transmission delays from causing a rate limit error to parse.

However, the same throttling issue persists when requesting. Contacting support to see if there is something I missed when reading the documentation about the traffic policies.

If it is something I am not directly missing, I will play around with some wait times between calls as it seems to throttle the connection after 40-50 requests.